+++ Changelog.md
@@ -4,6 +4,7 @@
 
 * Improve `--help` text
 * Improve error message when no tracks were found
+* Allow generated website to be placed in a non-root direcotory
 
 ## v0.1.0 - 2026-05-02