+++ Changelog.md
@@ -6,6 +6,7 @@
 * Improve error message when no tracks were found
 * Allow generated website to be placed in a non-root direcotory
 * Don't re-generate photos that are already present in the output directory
+* Accept multiple input directories
 
 ## v0.1.0 - 2026-05-02