+++ Changelog.md
@@ -2,11 +2,13 @@
 
 ## v0.1.1 - UNRELEASED
 
+* Plot elevation
 * Improve `--help` text
 * Improve error message when no tracks were found
 * Allow generated website to be placed in a non-root direcotory
 * Don't re-generate photos that are already present in the output directory
 * Accept multiple input directories
+* Show more information during the generator process (nr. photos per day, nr. skipped tiles)
 
 ## v0.1.0 - 2026-05-02