+++ Changelog.md
@@ -5,6 +5,7 @@
 * Plot elevation
 * Improve `--help` text
 * Improve error message when no tracks were found
+* Add GeoJSON support
 * Allow generated website to be placed in a non-root direcotory
 * Don't re-generate photos that are already present in the output directory
 * Accept multiple input paths which can now be either directories or files